Move Hive related suite into hive package
